% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_codes_state-data.R
\docType{data}
\name{fips_codes_state}
\alias{fips_codes_state}
\title{US FIPS Codes}
\format{
A data frame with 57 observations (50 US States, D.C., 5 territories, and minor outlying islands) and 3 variables:
\describe{
\item{state}{2-letter US Postal Service Code}
\item{state_code}{FIPS code}
\item{state_name}{Full state name}

}
}
\source{
\url{https://walker-data.com/tidycensus/}
}
\usage{
data(fips_codes_state)
}
\description{
Data set with 2-letter US postal service state code, FIPS code, and state name for easier
identification of 2-letter US postal service code.
}
\examples{
data(fips_codes_state)
head(fips_codes_state$state)
}
\references{
Walker K, Herman M (2023). tidycensus: Load US Census Boundary and Attribute Data as 'tidyverse' and 'sf'-Ready Data Frames. R package version 1.4.1, https://walker-data.com/tidycensus/.
}
\keyword{datasets}
