% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map1_categorical_county.R
\name{map1_categorical_county}
\alias{map1_categorical_county}
\title{Categorical choropleth US map with territory geometries at the County Level}
\usage{
map1_categorical_county(
  data,
  join_var,
  county_data_year = "2020",
  fill_var,
  fill_color,
  fill_linewidth = 0.5,
  fill_linecolor = "gray50",
  legend_name = NULL,
  legend_face = "plain",
  inset_box_color = "black",
  territory_label_color = "black",
  title = "",
  border_ids = NULL,
  border_color = NULL,
  border_linewidth = 1,
  state_color = "black",
  state_linewidth = 1,
  save.filepath
)
}
\arguments{
\item{data}{Data frame that already includes the fill variable as a factor. See example code/vignette for more detail.}

\item{join_var}{Variable to join. Must be 5 number GEOID formatted as a character variable. See example code/vignette for more detail.}

\item{county_data_year}{Package data frame of county geometry. Two options: "2010" for 2010 Counties and "2020 for 2020 counties. Default "2020".}

\item{fill_var}{Categorical mapping variable entered as "variable".}

\item{fill_color}{Values for \code{scale_fill_manual()}. Recommended to prepare with labels, see example.}

\item{fill_linewidth}{County geometry border line width. Default linewidth = 0.5.}

\item{fill_linecolor}{County geometry border line color. Default color = "gray50".}

\item{legend_name}{Legend title entered as a string.}

\item{legend_face}{Legend title face. Uses standard ggplot2 element_text face options (e.g.,"plain", "italic", "bold", "bold.italic").}

\item{inset_box_color}{Color of inset box for HI, AK, and territories. Set as "white" to remove border.}

\item{territory_label_color}{Color of territory labels. Set as "white" to remove label.}

\item{title}{Figure title entered as a string.}

\item{border_ids}{List of county GEOIDs for option to outline specific states and territories.}

\item{border_color}{Color of optional county highlight border.}

\item{border_linewidth}{Linewidth of optional county highlight border. Default linewidth = 1.}

\item{state_color}{Color of state outline. Default color = "black". Set to NULL to remove outline.}

\item{state_linewidth}{Linewidth of state outline. Default linewidth = 1.}

\item{save.filepath}{File path for saving plot as "path/image.png".}
}
\description{
This function allows for choropleth mapping of a predefined categorical variable
for territory geometries at the county level.
See \code{map1_categorical()} for choropleth mapping of territory geometries at the state level, and
\code{map2_categorical_county()} for choropleth mapping of territory labels at the county level.
}
\examples{
# Example 1 Using Census Insurance Data, 2020 County Geometry
colors.census <- c("Less than 5\%" = "#feebe2", 
                   "5\% to <10\%" = "#f768a1", 
                   "10\% or Greater" = "#7a0177")

border <- c("13031", "13089", "13121")

map1_categorical_county(data = census.uninsured19.co, 
                       join_var = "GEOID", 
                       fill_var = "Percent.Cat", 
                       fill_color = colors.census, 
                       fill_linewidth = 0.5, 
                       fill_linecolor = "gray50",
                       legend_name = "Percent Uninsured",
                       title = "Figure 1. Percent Uninsured, Ages <19 Years",
                       border_ids = border,
                       border_color = "red",
                       border_linewidth = 0.5,
                       state_color = "black", 
                       state_linewidth = 1,
                       save.filepath = "saved-maps/map1-test-co-highlight.png")

}
